// client/src/pages/store-details/types.ts
export interface StoreFormData {
  name: string;
  type: string;
  link: string;
  about: string;
  member_since: string;
  address: string;
  city: string;
  state: string;
  phone: string;
  website: string;
  email: string;
  followers: string;
  hours: string;
  vibe_score: string;
  sales_ready_summary: string;
}